Refactor solver and test_variants into submodules
