+++ operators/tests/__init__.py
@@ -1,0 +1,7 @@
+from . import pattern, motion, channel
+
+operator_classes = (
+    *pattern.operator_classes,
+    *motion.operator_classes,
+    *channel.operator_classes,
+)
