+++ operators/tests/pattern.py
@@ -4,10 +4,6 @@ from ..helpers.tracking_helpers import (
     evaluate_motion_models,
     evaluate_channel_combinations,
 )
-
-
-
-
 class CLIP_OT_test_pattern(bpy.types.Operator):
     bl_idname = "clip.test_pattern"
     bl_label = "Test Pattern"
@@ -65,55 +61,6 @@ class CLIP_OT_test_pattern(bpy.types.Operator):
         return {'FINISHED'}
 
 
-class CLIP_OT_test_motion(bpy.types.Operator):
-    bl_idname = "clip.test_motion"
-    bl_label = "Test Motion"
-    bl_description = "Testet verschiedene Motion Models"
 
-    def execute(self, context):
-        clip = context.space_data.clip
-        if not clip:
-            self.report({'WARNING'}, "Kein Clip geladen")
-            return {'CANCELLED'}
-        motion_models = ['Loc', 'LocRot', 'LocScale', 'LocRotScale', 'Affine', 'Perspective']
-        best_model, score, error_val = evaluate_motion_models(context, motion_models)
-        context.scene.test_value = motion_models.index(best_model)
-        print(
-            f"[Test Motion] best_model={best_model} frames={score} error={error_val:.4f}"
-        )
-        self.report({'INFO'}, f"Best Motion Model: {best_model}")
-        return {'FINISHED'}
+operator_classes = (CLIP_OT_test_pattern,)
 
-
-class CLIP_OT_test_channel(bpy.types.Operator):
-    bl_idname = "clip.test_channel"
-    bl_label = "Test Channel"
-    bl_description = "Testet verschiedene Farbkanal-Kombinationen"
-
-    def execute(self, context):
-        clip = context.space_data.clip
-        if not clip:
-            self.report({'WARNING'}, "Kein Clip geladen")
-            return {'CANCELLED'}
-        combos = [
-            (True, False, False),
-            (True, True, False),
-            (True, True, True),
-            (False, True, False),
-            (False, True, True),
-            (False, False, True),
-        ]
-        best_combo, score, error_val = evaluate_channel_combinations(context, combos)
-        context.scene.test_value = combos.index(best_combo)
-        print(
-            f"[Test Channel] best_combo={best_combo} frames={score} error={error_val:.4f}"
-        )
-        self.report({'INFO'}, "Beste Kanal-Einstellung gewählt")
-        return {'FINISHED'}
-
-
-operator_classes = (
-    CLIP_OT_test_pattern,
-    CLIP_OT_test_motion,
-    CLIP_OT_test_channel,
-)
